+++ ui/src/router/modules/2knowledge-shared-system.ts
@@ -6,12 +6,18 @@ const ModelRouter = {
   redirect: '/knowledge',
   component: () => import('@/layout/layout-template/SimpleLayout.vue'),
   children: [
-
     {
       path: '/knowledge/system/document/upload/shared',
       name: 'UploadDocumentSharedSystem',
       meta: { activeMenu: '/knowledge/system' },
       component: () => import('@/views/shared/document-shared/UploadDocument.vue'),
+      hidden: true,
+    },
+    {
+      path: '/knowledge/system/import/shared',
+      name: 'ImportLarkDocumentShared',
+      meta: { activeMenu: '/knowledge' },
+      component: () => import('@/views/shared/document-shared/ImportLarkDocument.vue'),
       hidden: true,
     },
   ],
